modularized test changes
moved negative test suite to import-neg-test.js, moved config test suite to config-import-test.js, moved all shared import test functions to import-helper.js.
