+++ index.js
@@ -1,1 +1,2 @@
 module.exports = require('./lib/import');
+
