+++ test/import-test.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const {expect} = require('chai');
 const {importFromFilePath, importConfigFromFilePath, importCIMCOREFromFilePath, setLogger} = require('../index');
+const {id, pid, expectAndGetElement, expectAndGetEntry, expectAndGetDataElement, expectValue, expectPrimitiveValue, expectRefValue, expectChoiceValue, expectMinMax, expectCardOne, expectChoiceOption, expectField, expectConcept, expectIdentifier, expectPrimitiveIdentifier, expectNoConstraints, importFixture, importFixtureFolder, importConfiguration, importConfigurationFolder, importCimcoreNSFile, importCimcoreDEFile, importCimcoreVSFile, importCimcoreMapFile, importCimcoreProjectFile, importCimcoreFolder, checkImportErrors, convertSpecsToCimcore } = require('../test/import-helper');
 const {Version, DataElement, Value, RefValue, ChoiceValue, IncompleteValue, Identifier, PrimitiveIdentifier, Cardinality, ValueSetConstraint, CodeConstraint, IncludesCodeConstraint, BooleanConstraint, TypeConstraint, CardConstraint, TBD, REQUIRED, EXTENSIBLE, PREFERRED, EXAMPLE} = require('shr-models');
 const err = require('shr-test-helpers/errors');
 
@@ -135,7 +136,8 @@ describe('#importDataElement', () => {
 
   it('Import11: should correctly import a group entry with both a value and properties', () => {
     const specifications = importFixture('GroupValueAndProperties');
-    const group = expectAndGetEntry(specifications,'shr.test', 'GroupValueAndProperties');
+    // const group = expectAndGetEntry(specifications,'shr.test', 'GroupValueAndProperties');
+    const group = expectAndGetElement(specifications,'shr.test', 'GroupValueAndProperties');
     expectCardOne(group.value);
     expectPrimitiveValue(group.value, 'concept');
     expectNoConstraints(group.value);
@@ -759,222 +761,9 @@ describe('#importDataElement', () => {
   });
 });
 
-describe('#importDataElementCounterexample', () => {
-  beforeEach(function() {
-    err.clear();
-  });
-  
-  it('Neg1: should return errors when there are invalid vocabulary references', () => {
-    const specifications = importFixture('InvalidVocabularyReference', 1);
-    const simple = expectAndGetEntry(specifications, 'shr.test', 'Simple');
-    expect(simple.concepts).to.have.length(1);
-    expectConcept(simple.concepts[0], 'ZOO', 'bear'); // Defaults to vocabulary alias
-    expect(simple.description).to.equal('It is a simple entry with invalid vocab');
-    expectCardOne(simple.value);
-    expectPrimitiveValue(simple.value, 'string');
-    expectNoConstraints(simple.value);
-  });
+// mlt: modularized negative tests. Now located in import-neg-test.js
+// mlt: modularized config tests. Now located in config-import-test.js
 
-  it('Neg2: should return errors when there are invalid element references', () => {
-    const specifications = importFixture('InvalidElementReference', 1);
-    const simple = expectAndGetEntry(specifications, 'shr.test', 'Simple');
-    expect(simple.concepts).to.have.length(1);
-    expectConcept(simple.concepts[0], 'http://foo.org', 'bar');
-    expect(simple.description).to.equal('It is a simple entry with invalid element reference');
-    expectCardOne(simple.value);
-    expectValue(simple.value, 'unknown', 'Complex');
-    expectNoConstraints(simple.value);
-  });
-
-  it('Neg3: should return errors when there are invalid fully qualified element references', () => {
-    const specifications = importFixture('InvalidFQElementReference', 1);
-    const simple = expectAndGetEntry(specifications, 'shr.test', 'Simple');
-    expect(simple.concepts).to.have.length(1);
-    expectConcept(simple.concepts[0], 'http://foo.org', 'bar');
-    expect(simple.description).to.equal('It is a simple entry with invalid fully qualified element reference');
-    expectCardOne(simple.value);
-    expectValue(simple.value, 'other.ns', 'Complex');
-    expectNoConstraints(simple.value);
-  });
-
-  it('Neg4: should return errors when there are ambiguous element references', () => {
-    const specifications = importFixtureFolder('ambiguousResolution', 1);
-    const amb = expectAndGetEntry(specifications, 'shr.test.one', 'Ambiguous');
-    expect(amb.concepts).to.be.empty;
-    expect(amb.description).to.equal('It is an entry that uses an ambiguous reference');
-    expectCardOne(amb.value);
-    expectValue(amb.value, 'shr.test.two', 'Foo'); // Defaults to first encountered namespace
-    expectNoConstraints(amb.value);
-  });
-
-  it('Neg5: should return errors when there are conflicting vocab references', () => {
-    const specifications = importFixtureFolder('conflictingVocab', 1);
-    expect(err.errors()[0].msg).to.contain('FOO');
-    expect(err.errors()[0].msg).to.not.contain('MOO');
-
-    const conflicting = expectAndGetEntry(specifications, 'shr.test.one', 'Conflicting');
-    expect(conflicting.concepts).to.have.length(2);
-    expectConcept(conflicting.concepts[0], 'http://foo.org', 'bar'); // Default to the first encountered vocab
-    expectConcept(conflicting.concepts[1], 'http://moo.org', 'car');
-    expect(conflicting.description).to.equal('It is an entry that uses a conflicting vocab reference');
-    expectCardOne(conflicting.value);
-    expectPrimitiveValue(conflicting.value, 'string');
-    expectNoConstraints(conflicting.value);
-  });
-});
-
-
-describe('#importConfig', () => {
-  beforeEach(function() {
-    err.clear();
-  });
-
-  it('Config1: should correctly import a basic configuration', () => {
-    const configuration = importConfiguration('basicconfig');
-    expect(configuration).to.have.all.keys('projectName','projectShorthand','projectURL','provenanceInfo','publisher','contact','fhirURL','implementationGuide');
-    expect(configuration.projectName).to.eql('Test Project');
-    expect(configuration.projectShorthand).to.eql('TEST');
-    expect(configuration.projectURL).to.eql('http://test.org');
-    expect(configuration.fhirURL).to.eql('http://test.org/fhir');
-    expect(configuration.implementationGuide).to.eql({
-      'npmName': 'basic',
-      'version': '1.2.3',
-      'indexContent': 'basicindexcontent.html'
-    });
-    expect(configuration.publisher).to.eql('Test Publisher');
-    expect(configuration.contact).to.eql([{
-      'telecom': [{
-        'system': 'url',
-        'value': 'http://test.org'
-      }]
-    }]);
-  });
-
-  it('Config2: should correctly generate missing fhir url from project url when fhir url is missing', () => {
-    const configuration = importConfiguration('incompletefhirconfig');
-    expect(configuration).to.have.all.keys('projectName','projectShorthand','projectURL','provenanceInfo','publisher','contact','fhirURL','implementationGuide');
-    expect(configuration.projectName).to.eql('Test Project');
-    expect(configuration.projectShorthand).to.eql('TEST');
-    expect(configuration.projectURL).to.eql('http://test.org');
-    expect(configuration.fhirURL).to.eql('http://test.org/fhir');
-    expect(configuration.implementationGuide).to.eql({
-      'npmName': 'basic',
-      'version': '1.2.3',
-      'indexContent': 'basicindexcontent.html'
-    });
-    expect(configuration.publisher).to.eql('Test Publisher');
-    expect(configuration.contact).to.eql([{
-      'telecom': [{
-        'system': 'url',
-        'value': 'http://test.org'
-      }]
-    }]);
-
-  });
-
-  it('Config3: should correctly import an incomplete configuration with partial default data', () => {
-    const configuration = importConfiguration('incompleteconfig');
-    expect(configuration).to.have.all.keys('projectName','projectShorthand','projectURL','provenanceInfo','publisher','contact','fhirURL','implementationGuide');
-    expect(configuration.projectName).to.eql('Test Project');
-    expect(configuration.projectShorthand).to.eql('EXAMPLE');
-    expect(configuration.projectURL).to.eql('http://example.com');
-    expect(configuration.fhirURL).to.eql('http://example.com/fhir');
-    expect(configuration.implementationGuide).to.eql({
-      'npmName': 'example-npm-name',
-      'version': '0.0.1',
-      'indexContent': 'exampleIndexContent.html'
-    });
-    expect(configuration.publisher).to.eql('Example Publisher');
-    expect(configuration.contact).to.eql([{
-      'telecom': [{
-        'system': 'url',
-        'value': 'http://test.org'
-      }]
-    }]);
-  });
-
-  it('Config4: should correctly import a configuration with missing npmName & version and add values', () => {
-    const configuration = importConfiguration('incompleteigconfig');
-    expect(configuration).to.have.all.keys('projectName','projectShorthand','projectURL','provenanceInfo','publisher','contact','fhirURL','implementationGuide');
-    expect(configuration.projectName).to.eql('Test Project');
-    expect(configuration.projectShorthand).to.eql('TEST');
-    expect(configuration.projectURL).to.eql('http://test.org');
-    expect(configuration.fhirURL).to.eql('http://test.org/fhir');
-    expect(configuration.implementationGuide).to.eql({
-      'npmName': 'example-npm-name',
-      'version': '0.0.1',
-      'indexContent': 'basicindexcontent.html'
-    });
-    expect(configuration.publisher).to.eql('Test Publisher');
-    expect(configuration.contact).to.eql([{
-      'telecom': [{
-        'system': 'url',
-        'value': 'http://test.org'
-      }]
-    }]);
-  });
-
-  it('Config5: should correctly translate deprecated properties', () => {
-    const configuration = importConfiguration('deprecatedigconfig');
-    expect(configuration).to.have.all.keys('projectName','projectShorthand','projectURL','provenanceInfo','publisher','contact','fhirURL','implementationGuide');
-    expect(configuration.projectName).to.eql('Test Project');
-    expect(configuration.projectShorthand).to.eql('TEST');
-    expect(configuration.projectURL).to.eql('http://test.org');
-    expect(configuration.fhirURL).to.eql('http://test.org/fhir');
-    expect(configuration.implementationGuide).to.eql({
-      'npmName': 'example-npm-name',
-      'version': '0.0.1',
-      'indexContent': 'basicindexcontent.html',
-      'includeLogicalModels': false,
-      'includeModelDoc': false,
-      'primarySelectionStrategy': {
-        'strategy': 'hybrid',
-        'primary': ['test'],
-        'hideSupporting': true
-      }
-    });
-    expect(configuration.publisher).to.eql('Test Publisher');
-    expect(configuration.contact).to.eql([{
-      'telecom': [{
-        'system': 'url',
-        'value': 'http://test.org'
-      }]
-    }]);
-  });
-
-  it('Config6: should correctly throw error when file is not valid JSON', () => {
-    const configuration = importConfiguration('invalidblankconfig', 1);
-    expect(err.errors()[0].msg).to.contain('Invalid config file');
-    expect(configuration).to.be.undefined;
-  });
-
-  it('Config7: should correctly create default configuration as config.json when no file exists', () => {
-    // First ensure config.json doesn't exist (from previous run)
-    const cfgPath = `${__dirname}/fixtures/config/emptyfolder/config.json`;
-    if (fs.existsSync(cfgPath)) {
-      fs.unlinkSync(cfgPath);
-    }
-    const configuration = importConfigurationFolder('emptyfolder');
-    expect(fs.existsSync(cfgPath)).to.be.true;
-    expect(configuration).to.have.all.keys('projectName','projectShorthand','projectURL','provenanceInfo','publisher','contact','fhirURL','implementationGuide');
-    expect(configuration.projectName).to.eql('Example Project');
-    expect(configuration.projectShorthand).to.eql('EXAMPLE');
-    expect(configuration.projectURL).to.eql('http://example.com');
-    expect(configuration.fhirURL).to.eql('http://example.com/fhir');
-    expect(configuration.implementationGuide).to.eql({
-      'npmName': 'example-npm-name',
-      'version': '0.0.1',
-      'indexContent': 'exampleIndexContent.html'
-    });
-    expect(configuration.publisher).to.eql('Example Publisher');
-    expect(configuration.contact).to.eql([{
-      'telecom': [{
-        'system': 'url',
-        'value': 'http://example.com'
-      }]
-    }]);
-  });
-});
 
 /* MK - cutting out CIMCORE tests because the examples are out of date, in Grammar 5. If we revive this, it should live in a separate file (e.g., cimcore-test.js).
 
@@ -1022,233 +811,4 @@ describe('#importCimcoreFromFilePath', () => {
 });
 */
 
-// Shorthand Identifier constructor for more concise code
-function id(namespace, name) {
-  return new Identifier(namespace, name);
-}
-
-// Shorthand PrimitiveIdentifier constructor for more concise code
-function pid(name) {
-  return new PrimitiveIdentifier(name);
-}
-
-function expectAndGetElement(specs, expectedNamespace, expectedName) {
-  return expectAndGetDataElement(specs, expectedNamespace, expectedName, false);
-}
-
-function expectAndGetEntry(specs, expectedNamespace, expectedName) {
-  return expectAndGetDataElement(specs, expectedNamespace, expectedName, true);
-}
-
-function expectAndGetDataElement(specs, expectedNamespace, expectedName, isEntry) {
-  const def = specs.dataElements.find(expectedNamespace, expectedName);
-  expect(def).to.be.instanceof(DataElement);
-  expect(def.isEntry).to.equal(isEntry);
-  expectIdentifier(def.identifier, expectedNamespace, expectedName);
-  return def;
-}
-
-function expectValue(value, expectedNamespace, expectedName) {
-  expect(value).to.be.instanceof(Value);
-  expectIdentifier(value.identifier, expectedNamespace, expectedName);
-}
-
-function expectPrimitiveValue(value, expectedName) {
-  expect(value).to.be.instanceof(Value);
-  expectPrimitiveIdentifier(value.identifier, expectedName);
-}
-
-function expectRefValue(value, expectedNamespace, expectedName) {
-  expect(value).to.be.instanceof(RefValue);
-  expectIdentifier(value.identifier, expectedNamespace, expectedName);
-}
-
-function expectChoiceValue(value, size) {
-  expect(value).to.be.instanceof(ChoiceValue);
-  expect(value.options).to.have.length(size);
-}
-
-function expectMinMax(value, expectedMin, expectedMax) {
-  expect(value).to.be.instanceof(Value);
-  const card = value.card;
-  if (typeof expectedMax == 'undefined') {
-    expectedMax = expectedMin;
-  }
-  if (typeof expectedMin !== 'undefined') {
-    expect(card).to.be.instanceof(Cardinality);
-    expect(card.min).to.equal(expectedMin);
-    if (typeof card.max !== 'undefined') {
-      expect(card.max).to.equal(expectedMax);
-      expect(card.isMaxUnbounded).to.be.false;
-    } else {
-      expect(card.max).to.be.undefined;
-      expect(card.isMaxUnbounded).to.be.true;
-    }
-  } else {
-    expect(card).to.be.undefined;
-  }
-}
-
-function expectCardOne(value) {
-  expect(value.card.isExactlyOne).to.be.true;
-}
-
-function expectChoiceOption(choice, optionIndex, expectedNamespace, expectedName, expectedMin=1, expectedMax=1) {
-  let option = choice.options[optionIndex];
-  expectMinMax(option, expectedMin, expectedMax);
-  expectValue(option, expectedNamespace, expectedName);
-}
-
-function expectField(element, fieldIndex, expectedNamespace, expectedName, expectedMin, expectedMax) {
-  const sptEl = element.fields[fieldIndex];
-  expectMinMax(sptEl, expectedMin, expectedMax);
-  expectValue(sptEl, expectedNamespace, expectedName);
-}
-
-function expectConcept(concept, system, code, display) {
-  expect(concept.system).equals(system);
-  expect(concept.code).equals(code);
-  expect(concept.display).equals(display);
-}
-
-function expectIdentifier(identifier, expectedNamespace, expectedName) {
-  expect(identifier).to.be.instanceof(Identifier);
-  expect(identifier.namespace).to.equal(expectedNamespace);
-  expect(identifier.name).to.equal(expectedName);
-}
-
-function expectPrimitiveIdentifier(identifier, expectedName) {
-  expect(identifier).to.be.instanceof(PrimitiveIdentifier);
-  expect(identifier.namespace).to.equal('primitive');
-  expect(identifier.name).to.equal(expectedName);
-}
-
-function expectNoConstraints(value) {
-  if (Array.isArray(value)) {
-    for (const v of value) {
-      expectNoConstraints(v);
-    }
-  } else {
-    expect(value.hasConstraints).to.be.false;
-  }
-}
-
-function importFixture(name, numExpectedErrors = 0) {
-  const dependencies = importFromFilePath(`${__dirname}/fixtures/dataElement/_dependencies`);
-  const specifications = importFromFilePath(`${__dirname}/fixtures/dataElement/${name}.txt`, null, dependencies);
-  checkImportErrors(numExpectedErrors);
-  return specifications;
-}
-
-function importFixtureFolder(name, numExpectedErrors = 0) {
-  const dependencies = importFromFilePath(`${__dirname}/fixtures/dataElement/_dependencies`);
-  const specifications = importFromFilePath(`${__dirname}/fixtures/dataElement/${name}`, null, dependencies);
-  checkImportErrors(numExpectedErrors);
-  return specifications;
-}
-
-function importConfiguration(name, numExpectedErrors = 0) {
-  const configuration = importConfigFromFilePath(`${__dirname}/fixtures/config/${name}.txt`);
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-}
-
-function importConfigurationFolder(name, numExpectedErrors = 0) {
-  const configuration = importConfigFromFilePath(`${__dirname}/fixtures/config/${name}`);
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-}
-
-function importCimcoreNSFile(namespace, numExpectedErrors = 0) {
-  namespace = namespace.replace(/\./g,'-');
-  const configuration = fs.readFileSync(`${__dirname}/fixtures/cimcore/${namespace}/${namespace}.json`, 'utf8');
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-}
-
-function importCimcoreDEFile(namespace, name, numExpectedErrors = 0) {
-  namespace = namespace.replace(/\./g,'-');
-  const configuration = fs.readFileSync(`${__dirname}/fixtures/cimcore/${namespace}/${namespace}-${name}.json`, 'utf8');
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-}
-
-function importCimcoreVSFile(namespace, name, numExpectedErrors = 0) {
-  namespace = namespace.replace(/\./g,'-');
-  const configuration = fs.readFileSync(`${__dirname}/fixtures/cimcore/${namespace}/valuesets/${name}.json`, 'utf8');
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-}
-
-function importCimcoreMapFile(namespace, name, numExpectedErrors = 0) {
-  namespace = namespace.replace(/\./g,'-');
-  const configuration = fs.readFileSync(`${__dirname}/fixtures/cimcore/${namespace}/mappings/${name}-mapping.json`, 'utf8');
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-
-}
-
-function importCimcoreProjectFile(numExpectedErrors = 0) {
-  const configuration = fs.readFileSync(`${__dirname}/fixtures/cimcore/project.json`, 'utf8');
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-}
-
-function importCimcoreFolder(numExpectedErrors = 0) {
-  const configuration = importCIMCOREFromFilePath(`${__dirname}/fixtures/cimcore/`);
-  checkImportErrors(numExpectedErrors);
-  return configuration;
-}
-
-function checkImportErrors(numExpectedErrors = 0) {
-  const errors = err.errors();
-  const message = `Import Errors: ${errors.map(e => e.msg).join('; ')}`;
-  expect(errors.length, message).to.equal(numExpectedErrors);
-}
-
-//THIS IS A HORRIBLE HACK. THIS AND THE CORRESPONDING SECTION IN shr-cli SHOULD
-//BE REPLACED WITH A STANDARD shr-model METHOD AS SOON AS WE REFACTOR shr-models.
-function convertSpecsToCimcore(configSpecifications, expSpecifications) {
-  const cimcoreSpecifications = {
-    'dataElements': [],
-    'valueSets': [],
-    'mappings': [],
-    'namespaces': {},
-  //also includes 'projectInfo'
-  };
-
-  //meta project file
-  let versionInfo = {
-    'CIMPL_version': '5.6.0',   //TODO: Update these to use an accurate constant. Current mode constant is on 4.0.0
-    'CIMCORE_version': '1.1'    //TODO: Update these to use an accurate constant. Current mode constant is on 4.0.0
-  };
-
-  let projectMetaOutput = Object.assign({ 'fileType': 'ProjectInfo' }, configSpecifications, versionInfo); //project meta information
-  cimcoreSpecifications['projectInfo'] = projectMetaOutput;
-
-  //meta namespace files
-  for (const ns of expSpecifications.namespaces.all) { //namespace files
-    let out = Object.assign({ 'fileType': 'Namespace' }, ns.toJSON());
-    cimcoreSpecifications.namespaces[ns.namespace] = out;
-  }
-
-  //data elements
-  for (const de of expSpecifications.dataElements.all) {
-    let out = Object.assign({ 'fileType': 'DataElement' }, de.toJSON());
-    cimcoreSpecifications.dataElements.push(out);
-  }
-
-  //valuesets
-  for (const vs of expSpecifications.valueSets.all) {
-    let out = Object.assign({ 'fileType': 'ValueSet' }, vs.toJSON());
-    cimcoreSpecifications.valueSets.push(out);
-  }
-
-  //mappings
-  for (const mapping of [...expSpecifications.maps._targetMap][0][1].all) {
-    let out = Object.assign({ 'fileType': 'Mapping' }, mapping.toJSON());
-    cimcoreSpecifications.mappings.push(out);
-  }
-
-  return cimcoreSpecifications;
-}
+// mlt: decoupled all shared function tests. Now located in import-helper.js
